Functions: `inline` -> `static inline`
Define functions in the header-only include library as `static inline`.  This
breaks the C99-style preference for `inline` functions in favor of
usability.  In the C99 style, each translation unit needs to contain an
`extern inline` declaration for each used function *exactly once*, which
can be frustrating to deal with when nesting ctimer inside projects.  With
`static inline` there is no such need, but it is not safe to
compare *pointers* to the inline functions.